relation: traits: Refactor traits into separate file
