+++ plonk/src/circuit/transcript.rs
@@ -21,7 +21,8 @@ use mpc_relation::{
         ecc::{PointVariable, SWToTEConParam},
         ultraplonk::mod_arith::FpElemVar,
     },
-    ConstraintSystem, PlonkCircuit, Variable,
+    traits::*,
+    PlonkCircuit, Variable,
 };
 
 /// Struct of variables representing a Rescue transcript type, including
@@ -233,7 +234,7 @@ mod tests {
     use ark_std::{format, UniformRand};
     use jf_primitives::pcs::prelude::{Commitment, UnivariateVerifierParam};
     use jf_utils::{bytes_to_field_elements, field_switching, test_rng};
-    use mpc_relation::{gadgets::ecc::TEPoint, Circuit};
+    use mpc_relation::gadgets::ecc::TEPoint;
 
     const RANGE_BIT_LEN_FOR_TEST: usize = 16;
     #[test]
