+++ plonk/src/proof_system/snark.rs
@@ -39,7 +39,7 @@ use jf_primitives::{
 };
 use jf_utils::par_utils::parallelizable_slice_iter;
 use mpc_relation::{
-    constants::compute_coset_representatives, gadgets::ecc::SWToTEConParam, Arithmetization,
+    constants::compute_coset_representatives, gadgets::ecc::SWToTEConParam, traits::*,
 };
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;
@@ -690,8 +690,8 @@ pub mod test {
     };
     use jf_utils::test_rng;
     use mpc_relation::{
-        constants::GATE_WIDTH, gadgets::ecc::SWToTEConParam, Arithmetization, Circuit,
-        ConstraintSystem, MergeableCircuitType, PlonkCircuit,
+        constants::GATE_WIDTH, gadgets::ecc::SWToTEConParam, traits::*, MergeableCircuitType,
+        PlonkCircuit,
     };
 
     // Different `m`s lead to different circuits.
