+++ primitives/src/circuit/commitment.rs
@@ -11,7 +11,7 @@ use crate::{
     utils::pad_with,
 };
 use ark_std::vec;
-use mpc_relation::{errors::CircuitError, ConstraintSystem, PlonkCircuit, Variable};
+use mpc_relation::{errors::CircuitError, traits::*, PlonkCircuit, Variable};
 
 use super::rescue::RescueNativeGadget;
 
@@ -53,7 +53,7 @@ mod tests {
     use ark_ff::UniformRand;
     use ark_std::vec::Vec;
     use itertools::Itertools;
-    use mpc_relation::{Circuit, PlonkCircuit, Variable};
+    use mpc_relation::{traits::*, PlonkCircuit, Variable};
 
     const TEST_INPUT_LEN: usize = 10;
     const TEST_INPUT_LEN_PLUS_ONE: usize = 10 + 1;
