+++ primitives/src/circuit/merkle_tree/universal_merkle_tree.rs
@@ -15,9 +15,7 @@ use crate::{
     rescue::RescueParameter,
 };
 use ark_std::vec::Vec;
-use mpc_relation::{
-    errors::CircuitError, BoolVar, Circuit, ConstraintSystem, PlonkCircuit, Variable,
-};
+use mpc_relation::{errors::CircuitError, traits::*, BoolVar, PlonkCircuit, Variable};
 
 type SparseMerkleTree<F> = RescueSparseMerkleTree<BigUint, F>;
 use num_bigint::BigUint;
@@ -134,7 +132,7 @@ mod test {
     use ark_ed_on_bls12_381_bandersnatch::Fq as FqEd381b;
     use ark_ed_on_bn254::Fq as FqEd254;
     use hashbrown::HashMap;
-    use mpc_relation::{Circuit, PlonkCircuit};
+    use mpc_relation::{traits::*, PlonkCircuit};
     use num_bigint::BigUint;
 
     type SparseMerkleTree<F> = RescueSparseMerkleTree<BigUint, F>;
