+++ relation/src/gadgets/mod.rs
@@ -22,7 +22,7 @@ pub use range::*;
 
 /// Utils for test
 pub mod test_utils {
-    use crate::{errors::CircuitError, Arithmetization, Circuit, PlonkCircuit};
+    use crate::{errors::CircuitError, traits::*, PlonkCircuit};
     use ark_ff::PrimeField;
 
     /// two circuit with the same statement should have the same extended
