+++ relation/src/gadgets/ultraplonk/mod_arith.rs
@@ -8,7 +8,9 @@
 use crate::{
     constants::GATE_WIDTH,
     errors::CircuitError::{self, ParameterError},
-    next_multiple, Circuit, ConstraintSystem, PlonkCircuit, Variable,
+    next_multiple,
+    traits::*,
+    PlonkCircuit, Variable,
 };
 use ark_ff::PrimeField;
 use ark_std::{format, string::ToString, vec, vec::Vec};
