+++ relation/src/gadgets/ultraplonk/range.rs
@@ -7,7 +7,8 @@
 //! Range proof gates.
 use crate::{
     errors::CircuitError::{self, ParameterError},
-    Circuit, PlonkCircuit, Variable,
+    traits::*,
+    PlonkCircuit, Variable,
 };
 use ark_ff::{BigInteger, PrimeField};
 use ark_std::{string::ToString, vec::Vec};
@@ -87,14 +88,13 @@ fn decompose_le<F: PrimeField>(val: F, len: usize, range_bit_len: usize) -> Vec<
 #[cfg(test)]
 mod test {
     use super::*;
+    use crate::traits::ConstraintSystem;
     use ark_bls12_377::Fq as Fq377;
     use ark_ed_on_bls12_377::Fq as FqEd377;
     use ark_ed_on_bls12_381::Fq as FqEd381;
     use ark_ed_on_bn254::Fq as FqEd254;
     use ark_std::rand::Rng;
     use jf_utils::test_rng;
-
-    use crate::ConstraintSystem;
 
     const RANGE_BIT_LEN_FOR_TEST: usize = 8;
     const RANGE_SIZE_FOR_TEST: usize = 256;
