+++ scripts/run_tests.sh
@@ -6,5 +6,5 @@ export RUSTFLAGS="-C overflow-checks=on"
 
 cargo +nightly test --release -p jf-utils -- -Zunstable-options --report-time
 cargo +nightly test --release -p mpc-plonk --lib --bins --all-features -- -Zunstable-options --report-time
-cargo +nightly test --release -p jf-primitives --features test-srs -- -Zunstable-options --report-time # enable test-srs feature for gen_srs_for_testing
-cargo +nightly test --release -p mpc-relation -- -Zunstable-options --report-time
+cargo +nightly test --release -p jf-primitives --all-features -- -Zunstable-options --report-time # enable test-srs feature for gen_srs_for_testing
+cargo +nightly test --release -p mpc-relation --all-features -- -Zunstable-options --report-time
